将stack和local var改成宽度为1bit (#38)
* [mmzk] (fix) 修复native函数的param count和local var count不一致导致segfault

* [mmzk] (feat) 提前function中每个local variable的offset

* [mmzk] (feat) 1bit宽度的stack和local varaibles
1. global variables依然是统一8bit宽度
2. 暂时关闭了deepmem（使用自带malloc），因为会有segfault。具体问题暂时未知
3. 还没有处理好drop和select等需要知道stack顶端variable宽度的指令（目前都是统一读最上面4bit），可能需要引入额外结构来存储stack上var的大小

* [mmzk] (refactor) 用DEBUG macro来选择是否展示deep_debug的信息

+++ include/deep_interp.h
@@ -21,16 +21,16 @@ typedef enum DEEPFrameType {
 typedef struct DEEPInterpFrame {
     struct DEEPInterpFrame *prev_func_frame; //指向前一个函数帧
     struct DEEPFunction *function; //当前函数实例
-    uint64_t *sp; //操作数栈指针
+    uint8_t *sp; //操作数栈指针
     DEEPFrameType type; //类型
-    uint64_t *local_vars; //局部变量
+    uint8_t *local_vars; //局部变量
 } DEEPInterpFrame;
 
 //操作数栈
 typedef struct DEEPStack {
     int32_t capacity;
-    uint64_t *sp;
-    uint64_t *sp_end;
+    uint8_t *sp;
+    uint8_t *sp_end;
 } DEEPStack;
 
 //控制栈
@@ -43,9 +43,9 @@ typedef struct DEEPControlStack {
 typedef struct DEEPExecEnv {
 
     struct DEEPInterpFrame *cur_frame; //当前函数帧
-    uint64_t *sp_end; //操作数栈大小
-    uint64_t *sp; //sp指针
-    uint64_t *local_vars; //函数局部变量
+    uint8_t *sp_end; //操作数栈大小
+    uint8_t *sp; //sp指针
+    uint8_t *local_vars; //函数局部变量
     uint64_t *global_vars; //全局变量
     uint8_t *memory; //内存
     DEEPControlStack *control_stack; //控制栈
